Finish setting up defillama mappings to BQ
