+++ helm/dagster/secret-provider.yaml
@@ -7,5 +7,5 @@ spec:
   provider: gke
   parameters:
     secrets: |
-      - resourceName: "projects/oplabs-tools-data/secrets/op-analytics-python-runner/versions/5"
+      - resourceName: "projects/oplabs-tools-data/secrets/op-analytics-python-runner/versions/6"
         path: "op-analytics-vault.txt"
