+++ src/op_analytics/datasources/defillama/mappings/token_mappings.py
@@ -1,10 +1,9 @@
 import polars as pl
 
-from op_analytics.coreutils.gsheets import read_gsheet
-from op_analytics.coreutils.misc import raise_for_schema_mismatch
 from op_analytics.coreutils.partitioned.dailydata import DEFAULT_DT
 
-from .dataaccess import DefiLlama
+from ..dataaccess import DefiLlama
+from .utils import load_data_from_gsheet
 
 
 DATA_MAPPINGS_GSHEET_NAME = "data_mappings"
@@ -20,7 +19,9 @@ TOKEN_MAPPINGS_SCHEMA = {
 
 def execute():
     df = load_data_from_gsheet(
-        DATA_MAPPINGS_GSHEET_NAME, TOKEN_MAPPINGS_WORKSHEET_NAME, TOKEN_MAPPINGS_SCHEMA
+        gsheet_name=DATA_MAPPINGS_GSHEET_NAME,
+        worksheet_name=TOKEN_MAPPINGS_WORKSHEET_NAME,
+        expected_schema=TOKEN_MAPPINGS_SCHEMA,
     )
 
     # Deduplicate
@@ -33,22 +34,3 @@ def execute():
     DefiLlama.TOKEN_MAPPINGS.write(dataframe=df.with_columns(dt=pl.lit(DEFAULT_DT)))
 
     return {TOKEN_MAPPINGS_WORKSHEET_NAME: len(df)}
-
-
-def load_data_from_gsheet(
-    gsheet_name: str, worksheet_name: str, expected_schema: dict
-) -> pl.DataFrame:
-    # Read from Google Sheets Input
-    raw_records = read_gsheet(
-        location_name=gsheet_name,
-        worksheet_name=worksheet_name,
-    )
-    raw_df = pl.DataFrame(raw_records, infer_schema_length=len(raw_records))
-
-    # Validate schema
-    raise_for_schema_mismatch(
-        actual_schema=raw_df.schema,
-        expected_schema=pl.Schema(expected_schema),
-    )
-
-    return raw_df
